refactor(point): controller 개별 폴더로 이동
