+++ src/point/point.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PointController } from './point.controller';
+import { PointController } from './controller/point.controller';
 import { PointService } from './point.service';
 import { DatabaseModule } from 'src/database/database.module';
 import { PointPolicy } from './policy/point.policy';
